refactoring + routing upd + etc
