+++ Buildlease/Buildlease/ClientApp/src/API.tsx
@@ -1,37 +1,37 @@
 import axios from "axios";
-import GetProductsRequest from "./components/Catalogue/Request/GetProductsRequest";
-import CategoryFilterView from "./components/Views/CategoryFilterView";
-import ProductView from "./components/Views/ProductView";
-import CategoryFullView from "./components/Views/CategoryFullView";
+import GetProductsRequest from "./components/catalog/requests/GetProductsRequest";
+import CategoryFilterView from "./components/views/CategoryFilterView";
+import ProductView from "./components/views/ProductView";
+import CategoryFullView from "./components/views/CategoryFullView";
 
 const MainLink = 'https://localhost:5001/api/';
 
 const API = {
 
-  GetAllCategories: () => {
-    return axios
-      .post<CategoryFullView[]>(MainLink + 'GetAllCategories')
-      .then(res => res.data);
-  },
-  
-  GetProducts: (info: GetProductsRequest) => {
-    return axios
-      .post<ProductView[]>(MainLink + 'GetProducts', info)
-      .then(res => res.data);
-  },
+    GetAllCategories: () => {
+        return axios
+            .post<CategoryFullView[]>(MainLink + 'GetAllCategories')
+            .then(res => res.data);
+    },
 
-  GetCategoryFilters: (categoryId: number) => {
-    return axios
-      .post<CategoryFilterView[]>(MainLink + `GetCategoryFilters/${categoryId}`)
-      .then(res => res.data);
-  },
+    GetProducts: (info: GetProductsRequest) => {
+        return axios
+            .post<ProductView[]>(MainLink + 'GetProducts', info)
+            .then(res => res.data);
+    },
 
-  SetProductOrderCount: (productId: number, count: number) => {
-    return axios
-      .post<void>(MainLink + `SetProductOrderCount/${productId}/${count}`)
-      .then(res => res.data);
-  },
-  
+    GetCategoryFilters: (categoryId: number) => {
+        return axios
+            .post<CategoryFilterView[]>(MainLink + `GetCategoryFilters/${categoryId}`)
+            .then(res => res.data);
+    },
+
+    SetProductOrderCount: (productId: number, count: number) => {
+        return axios
+            .post<void>(MainLink + `SetProductOrderCount/${productId}/${count}`)
+            .then(res => res.data);
+    },
+
 }
 
 export default API;
