--- Buildlease/Buildlease/ClientApp/src/components/Catalogue/Request/AttributeFilter.tsx
@@ -1,6 +1,0 @@
-export default interface AttributeFilter {
-  AttributeId: number,
-  ValueNumberLowerBound: number | null,
-  ValueNumberUpperBound: number | null,
-  ValueStringAllowed: string[] | null,
-}
