--- Buildlease/Buildlease/ClientApp/src/components/Catalogue/Request/GetProductsRequest.tsx
@@ -1,10 +1,0 @@
-import AttributeFilter from "./AttributeFilter";
-import SortRule from "./SortRule"
-
-export default interface GetProductsRequest {
-  CategoryId: number,
-  Filters: AttributeFilter[] | null,
-  OrderByRule: SortRule | null,
-  SkipCount: number,
-  TakeCount: number,
-}
