--- Buildlease/Buildlease/ClientApp/src/components/Catalogue/Request/SortRule.tsx
@@ -1,11 +1,0 @@
-enum SortRule {
-  Default = 'Default',
-  PriceAscending = 'PriceAscending',
-  PriceDescending = 'PriceDescending',
-  NameAscending = 'NameAscending',
-  NameDescending = 'NameDescending',
-  PopularityAscending = 'PopularityAscending',
-  PopularityDescending = 'PopularityDescending',
-}
-
-export default SortRule;
