+++ Buildlease/Buildlease/ClientApp/src/components/NotFound.tsx
@@ -1,4 +1,4 @@
-function NotFound() {
+export default function NotFound() {
     return (
         <>
             <h1>404 page not found.</h1>
@@ -6,5 +6,3 @@ function NotFound() {
         </>
     )
 }
-
-export default NotFound
