+++ Buildlease/Buildlease/ClientApp/src/components/cart/Cart.tsx
@@ -1,9 +1,7 @@
-import { SubHeader } from "../layout/SubHeader"
-
 function Cart() {
     return (
         <>
-            <SubHeader/>
+            Cart
         </>
     )
 }
