+++ Buildlease/Buildlease/ClientApp/src/components/catalog/requests/AttributeFilter.tsx
@@ -1,0 +1,6 @@
+export default interface AttributeFilter {
+    AttributeId: number,
+    ValueNumberLowerBound: number | null,
+    ValueNumberUpperBound: number | null,
+    ValueStringAllowed: string[] | null,
+}
