+++ Buildlease/Buildlease/ClientApp/src/components/catalog/requests/GetProductsRequest.tsx
@@ -1,0 +1,10 @@
+import AttributeFilter from "./AttributeFilter";
+import SortRule from "./SortRule"
+
+export default interface GetProductsRequest {
+    CategoryId: number,
+    Filters: AttributeFilter[] | null,
+    OrderByRule: SortRule | null,
+    SkipCount: number,
+    TakeCount: number,
+}
