+++ Buildlease/Buildlease/ClientApp/src/components/catalog/requests/SortRule.tsx
@@ -1,0 +1,11 @@
+enum SortRule {
+    Default = 'Default',
+    PriceAscending = 'PriceAscending',
+    PriceDescending = 'PriceDescending',
+    NameAscending = 'NameAscending',
+    NameDescending = 'NameDescending',
+    PopularityAscending = 'PopularityAscending',
+    PopularityDescending = 'PopularityDescending',
+}
+
+export default SortRule;
