+++ Buildlease/Buildlease/ClientApp/src/components/layout/SideMenu.tsx
@@ -1,0 +1,10 @@
+// routing
+import { Route, Switch } from 'react-router-dom';
+
+export default function SideMenu() {
+    return(
+        <>
+            Side Menu
+        </>
+    );
+}
