+++ Buildlease/Buildlease/ClientApp/src/components/layout/SubHeader.tsx
@@ -1,4 +1,3 @@
-import { Component } from 'react';
 import { Route } from 'react-router-dom';
 
 import PageState from './PageState';
@@ -6,15 +5,13 @@ import SortRule from './SortRule';
 
 import styles from './layout.module.css';
 
-export class SubHeader extends Component {
-    render() {
-        return (
-            <div className={`d-flex justify-content-between align-items-center ${styles.wideBar}`}>
-                <PageState/>
-                <Route path='/catalog'>
-                    <SortRule/>
-                </Route>
-            </div>
-        );
-    }
+export default function SubHeader() {
+    return (
+        <div className={`d-flex justify-content-between align-items-center ${styles.wideBar}`}>
+            <PageState/>
+            <Route path='/catalog'>
+                <SortRule/>
+            </Route>
+        </div>
+    );
 }
